# Conflicts:
#	yeb1/yeb-server/src/main/java/com/xxxx/server/controller/AdminController.java
#	yeb1/yeb-server/src/main/java/com/xxxx/server/controller/LoginController.java
#	yeb1/yeb-server/src/main/java/com/xxxx/server/mapper/MenuMapper.java
#	yeb1/yeb-server/src/main/java/com/xxxx/server/mapper/MenuRoleMapper.java
#	yeb1/yeb-server/src/main/java/com/xxxx/server/service/IAdminService.java
#	yeb1/yeb-server/src/main/java/com/xxxx/server/service/IMenuRoleService.java
#	yeb1/yeb-server/src/main/java/com/xxxx/server/service/IMenuService.java
#	yeb1/yeb-server/src/main/java/com/xxxx/server/service/impl/AdminServiceImpl.java
#	yeb1/yeb-server/src/main/java/com/xxxx/server/service/impl/MenuRoleServiceImpl.java

+++ yeb1/yeb-server/src/main/java/com/xxxx/server/controller/AdminController.java
@@ -2,14 +2,17 @@ package com.xxxx.server.controller;
 
 
 import com.xxxx.server.pojo.Admin;
+import com.xxxx.server.pojo.RespBean;
 import com.xxxx.server.pojo.Role;
 import com.xxxx.server.service.IAdminService;
+import com.xxxx.server.service.impl.RoleServiceImpl;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import com.xxxx.server.pojo.RespBean;
 import com.xxxx.server.service.impl.RoleServiceImpl;
+import org.springframework.boot.ansi.AnsiOutput;
 import org.springframework.web.bind.annotation.*;
 
 
@@ -37,22 +40,34 @@ public class AdminController {
     private RoleServiceImpl roleService;
 
 
-    @ApiOperation(value = "更新操作员")
-    @PutMapping("/")
-    public RespBean updateAdmin(@RequestBody Admin admin) {
-        if (adminService.updateById(admin)) {
-            return RespBean.success("更新成功！");
+    @ApiOperation(value = "获取所有操作员")
+    @GetMapping("/system/admin")
+    public List<Admin> getAllAdmins(String keywords) {
+        return adminService.getAllAdmins(keywords);
         }
-        return RespBean.error("更新失败！");
-    }
 
-    @ApiOperation(value = "删除操作员")
-    @DeleteMapping("/{id}")
-    public RespBean deleteAdmin(@PathVariable Integer id) {
-        if (adminService.removeById(id)) {
-            return RespBean.success("删除成功！");
+    @ApiOperation(value = "更新操作员")
+    @PutMapping("/system/admin")
+    public RespBean updateAdmin(@RequestBody Admin admin) {
+            if (adminService.updateById(admin)) {
+                return RespBean.success("更新成功！");
+            }
+            return RespBean.error("更新失败！");
         }
-        return RespBean.error("删除失败！");
+
+    /**
+     只有系统管理员才能删除操作员
+     */
+    @ApiOperation(value = "删除操作员")
+    @DeleteMapping("/system/admin/{id}")
+    public RespBean deleteAdmin(@PathVariable Integer id) {
+        if (id == 1) {
+            if (adminService.removeById(id)) {
+                return RespBean.success("删除成功！");
+            }
+            return RespBean.error("删除失败！");
+        }
+        return RespBean.error("非常抱歉，您没有删除权限");
     }
 
     /**
@@ -62,12 +77,20 @@ public class AdminController {
      */
 
     @ApiOperation(value = "获取所有角色")
+    @GetMapping("/system/admin/roles")
+    public List<Role>getRoles(){
     @GetMapping("/roles")
     public List<Role> getRoles() {
         return roleService.list();
     }
 
     @ApiOperation(value = "更新操作员的角色")
+    @PutMapping("/system/admin/role")
+    public RespBean updateAdminRole(Integer adminId,Integer[] rids) {
+        if ( rids != null) {
+            return adminService.updateAdminRole(adminId, rids);
+        }
+        return RespBean.error("操作员角色不能为空");
     @PutMapping("/role")
     public RespBean updateAdminRole(Integer adminId, Integer[] rids) {
         return adminService.updateAdminRole(adminId, rids);
@@ -77,6 +100,14 @@ public class AdminController {
     @GetMapping("/")
     public List<Admin> getAllAdmins(String keywords) {
         return adminService.getAllAdmins(keywords);
+    @ApiOperation(value = "根据用户名查询对象")
+    @GetMapping("admin/info")
+    public Admin quryAdminByName(Principal principal) {
+        Admin admin = adminService.quryAdminByName(principal.getName());
+        admin.setPassword(null);
+        List<Role> roles = adminService.quryRoles(admin.getId());
+        admin.setRoles(roles);
+        return admin;
     }
 }
